server: move QUIC implementations into server module
